+++ NEWS.md
@@ -1,4 +1,6 @@
 # devel
+- `with_connection()` function to automatically close R file connections.
+
 - `with_gctorture2` command to run code with gctorture2, useful for testing
   (#47).